window issue fixed
with the following code, the WelcomeWindow closes(actually hides but process runs in the background) and TeacherModule window opens with a new window. I want to totally close welcome window.
furthermore, in following code approach, when TeacherModule window is closed with window's upper-right corner close button, window closes but WelcomeWindow runs in the background.